backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
openai==1.55.3
httpx==0.28.1
